finalised login mvvm srtucture
