+++ free-tunes/Extensions/UIViewControllerExtension.swift
@@ -9,25 +9,15 @@ import UIKit
 
 extension UIViewController {
     
-    func getAuthenticator() -> AuthManager {
-        return AuthManager()
-    }
-    
     func displayErrorAlert(alertTitle: String,
                            alertMessage: String,
-                           alertActionTitle: String,
-                           alertDelegate: UIViewController) {
-        if alertTitle != "Empty login fields" {
+                           alertActionTitle: String) {
             let alert = UIAlertController(title: alertTitle,
                                           message: alertMessage,
                                           preferredStyle: UIAlertController.Style.alert)
             alert.addAction(UIAlertAction(title: alertActionTitle, style: UIAlertAction.Style.default, handler: nil))
-            alertDelegate.present(alert, animated: true, completion: nil)
+            self.present(alert, animated: true, completion: nil)
             return
-        }
-        else{
-            alertDelegate
-        }
     }
     
 }
